Add ToMarkdownService class
Also exposes new endpoint `supported` for use in the AI binding
